requirements.txt
absl-py==2.1.0
asgiref==3.7.2
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.1
cryptography==42.0.5
cvzone==1.6.1
cycler==0.12.1
Django==5.0.3
django-blob-storage==0.2.2
django-cors-headers==4.3.1
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==7.0.0
Faker==24.8.0
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
idna==3.6
isodate==0.6.1
jax==0.4.26
jaxlib==0.4.26
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mediapipe==0.10.9
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pillow==10.3.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
scipy==1.13.0
six==1.16.0
sounddevice==0.4.6
sqlparse==0.4.4
sympy==1.12
torch==2.2.2
typing_extensions==4.10.0
urllib3==2.2.1
